move encoders to their own file
