+++ src/main/scala/com/unstablebuild/slime/package.scala
@@ -1,0 +1,5 @@
+package com.unstablebuild
+
+package object slime extends Encoders {
+
+}
